updated version of source code

+++ src/BLAS3.h
@@ -12,9 +12,9 @@ void spmm_csr_v2 (csrType_local csr_mat, denseType dense_mat, denseType *res_mat
 
 //double * global_swap_zone is deprecated
 
-void spmm_csr_info_data_sep_CBCG (csrType_local csr_mat, denseType dense_mat_info, double * dataSrc, int dataDisp
+void spmm_csr_info_data_sep_CBCG (csrType_local csr_mat, denseType dense_mat_info, double * dataSrc, long dataDisp
         , denseType *res_mat, int myid, int numprocs);
-void spmm_csr_info_data_sep_BCBCG (csrType_local csr_mat, denseType dense_mat_info, double * dataSrc, int dataDisp
+void spmm_csr_info_data_sep_BCBCG (csrType_local csr_mat, denseType dense_mat_info, double * dataSrc, long dataDisp
         , denseType *res_mat, int myid, int numprocs);
 
 // res = alpha*mat1 + beta*mat2, TP, third place
@@ -24,18 +24,18 @@ void dense_mat_mat_add_TP(denseType mat1, denseType mat2, denseType output_mat, 
 // res[output_mat_disp ...] = alpha*mat1 + beta*mat2, TP, third place
 // note: one each processor, mat1 and mat2 should have the same shape
 // res matrix get the result, however, a displacement should be promised
-void dense_mat_mat_add_TP_targetDisp(denseType mat1, denseType mat2, denseType output_mat, int output_mat_disp
+void dense_mat_mat_add_TP_targetDisp(denseType mat1, denseType mat2, denseType output_mat, long output_mat_disp
                                     ,double alpha, double beta, int myid, int numprocs);
 
 // res[output_mat_disp ...] = alpha*mat1[disp1...] + beta*mat2[disp2...] + gamma*mat3[disp3...], TP, third place
 // length data items will be processed
 // note: one each processor, mat1 and mat2 and mat3 should have the same shape
 // res matrix get the result, however, a displacement should be promised
-void dense_array_mat_mat_mat_add_TP_disp(double* mat1Data, int mat1Disp
-                                             , double* mat2Data, int mat2Disp
-                                             , double* mat3Data, int mat3Disp
-                                             , double* output_mat, int output_mat_disp
-                                             , int length
+void dense_array_mat_mat_mat_add_TP_disp(double* mat1Data, long mat1Disp
+                                             , double* mat2Data, long mat2Disp
+                                             , double* mat3Data, long mat3Disp
+                                             , double* output_mat, long output_mat_disp
+                                             , long length
                                              , double alpha, double beta, double gamma
                                              ,int myid, int numprocs);
 
@@ -65,3 +65,4 @@ void MatTranposeMatMul (denseType *resMat, denseType mat, int myid, int numprocs
 
 //matL * matR
 void dense_distributedMatL_localMatR_Mul_v1 (denseType * res_mat, denseType matDistL, denseType matLocalR,int myid, int numprocs);
+
