+++ src/MatrixOperations.c
@@ -5,10 +5,10 @@
 double* local_dense_colum_2norm(denseType mat) {
     double * norm_array = (double *) calloc(mat.local_num_col, sizeof (double));
 
-    int idx;
+    long idx;
     for (idx = 0; idx < mat.local_num_col; idx++)
         norm_array[idx] = 0.0;
-    int idx_i, idx_j;
+    long idx_i, idx_j;
     for (idx_i = 0; idx_i < mat.local_num_row; idx_i++) {
         for (idx_j = 0; idx_j < mat.local_num_col; idx_j++) {
             norm_array[idx_j] += mat.data[idx_i * mat.local_num_col + idx_j]
@@ -25,14 +25,14 @@ void dense_entry_copy(denseType src, denseType target) {
     assert(src.local_num_row == target.local_num_row);
 #endif
 
-    int idx;
-    int total_num_ele = src.local_num_col * src.local_num_row;
+    long idx;
+    long total_num_ele = src.local_num_col * src.local_num_row;
     for (idx = 0; idx < total_num_ele; idx++) {
         target.data[idx] = src.data[idx];
     }
 }
 
-void generate_no_parallel_dense(denseType * mat, int num_row, int num_col) {
+void generate_no_parallel_dense(denseType * mat, long num_row, long num_col) {
     mat->global_num_col = num_col;
     mat->global_num_row = num_row;
 
@@ -51,10 +51,10 @@ void generate_no_parallel_dense(denseType * mat, int num_row, int num_col) {
 
 void dense_matrix_local_transpose_row_order(denseType inMat, denseType outMat) {
 
-    int num_col = inMat.local_num_col;
-    int num_row = inMat.local_num_row;
+    long num_col = inMat.local_num_col;
+    long num_row = inMat.local_num_row;
 
-    int rowIdx, colIdx;
+    long rowIdx, colIdx;
 
     for (rowIdx = 0; rowIdx < num_row; rowIdx++) {
         for (colIdx = 0; colIdx < num_col; colIdx++) {
@@ -67,23 +67,23 @@ void dense_matrix_local_transpose_row_order(denseType inMat, denseType outMat) {
 // the elements of inMat change to vector of size of blocksize, 
 // and each of the vectors are stored in row order in memory
 
-void dense_matrix_local_transpose_chunk_row_order(denseType inMat, denseType outMat, int blocksize) {
+void dense_matrix_local_transpose_chunk_row_order(denseType inMat, denseType outMat, long blocksize) {
 
-    int num_col = inMat.local_num_col;
-    int num_col_vec = inMat.local_num_col/blocksize;
+    long num_col = inMat.local_num_col;
+    long num_col_vec = inMat.local_num_col/blocksize;
 #ifdef ASSERTION_DEBUG
     assert (inMat.local_num_col%blocksize == 0);
 #endif
-    int num_row = inMat.local_num_row;
-    int num_col_output = num_row * blocksize;
+    long num_row = inMat.local_num_row;
+    long num_col_output = num_row * blocksize;
 
-    int rowIdx, colIdx;
-    int blockIdx;
+    long rowIdx, colIdx;
+    long blockIdx;
 
     for (rowIdx = 0; rowIdx < num_row; rowIdx++) {
         for (colIdx = 0; colIdx < num_col_vec; colIdx++) {
-            int chunkIdx_in=rowIdx * num_col + colIdx*blocksize;
-            int chunkIdx_out = colIdx * num_col_output + rowIdx*blocksize;
+            long chunkIdx_in=rowIdx * num_col + colIdx*blocksize;
+            long chunkIdx_out = colIdx * num_col_output + rowIdx*blocksize;
             for (blockIdx = 0; blockIdx < blocksize; blockIdx++) {
                 outMat.data[chunkIdx_out+blockIdx] = inMat.data[chunkIdx_in+blockIdx];
             }
@@ -92,7 +92,7 @@ void dense_matrix_local_transpose_chunk_row_order(denseType inMat, denseType out
 }
 
 void local_dense_mat_print(denseType mat, int myid) {
-    int rowIdx, colIdx;
+    long rowIdx, colIdx;
 
     for (rowIdx = 0; rowIdx < mat.local_num_row; rowIdx++) {
         printf("myid: %d, row: %d\n", myid, rowIdx);
@@ -107,12 +107,12 @@ void local_dense_mat_print(denseType mat, int myid) {
 
 //#define norm2square_dist_denseMat_col_n_DB
 
-void norm2square_dist_denseMat_col_n(denseType mat, double* result, int n, int myid, int numprocs) {
-    int ierr;
+void norm2square_dist_denseMat_col_n(denseType mat, double* result, long n, int myid, int numprocs) {
+    long ierr;
     double partial_result = .0;
-    int c_n = n - 1;
-    int rowIdx;
-    int col_num_length = mat.local_num_col;
+    long c_n = n - 1;
+    long rowIdx;
+    long col_num_length = mat.local_num_col;
     if (n > col_num_length) {
         exit(0);
     }
@@ -126,6 +126,9 @@ void norm2square_dist_denseMat_col_n(denseType mat, double* result, int n, int m
         printf("partial result:%f\n", partial_result);
     }
 #endif
+    
+// int MPI_Allreduce(const void *sendbuf, void *recvbuf, int count,
+//                   MPI_Datatype datatype, MPI_Op op, MPI_Comm comm)
 
     ierr = MPI_Allreduce((void*) &partial_result, (void*) result, 1
             , MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
@@ -138,17 +141,17 @@ void norm2square_dist_denseMat_col_n(denseType mat, double* result, int n, int m
 
 }
 
-void dense_entry_copy_disp(denseType src, int srcDispStart, denseType target, int tarDispStart, int count) {
+void dense_entry_copy_disp(denseType src, long srcDispStart, denseType target, long tarDispStart, long count) {
 #ifdef ASSERTION_DEBUG
-    int srcTotal = src.local_num_col * src.local_num_row;
-    int targetTotal = target.local_num_col * target.local_num_row;
+    long srcTotal = src.local_num_col * src.local_num_row;
+    long targetTotal = target.local_num_col * target.local_num_row;
 
     assert((srcDispStart + count) <= srcTotal);
     assert((tarDispStart + count) <= targetTotal);
 #endif
 
-    int srcIdx, tarIdx;
-    int srcEndIdx = srcDispStart + count;
+    long srcIdx, tarIdx;
+    long srcEndIdx = srcDispStart + count;
     tarIdx = tarDispStart;
     for (srcIdx = srcDispStart; srcIdx < srcEndIdx; srcIdx++) {
